requirements.txt
torch
opencv-python
numpy
pillow
torchvision